Moved to Maven project, made project structure
