+++ simple-search-engine-maven/src/main/resources/text.txt
